fixed: meal refactoring
