+++ src/components/meal/detail-meal/item/index.ts
@@ -1,0 +1,1 @@
+export { default } from './detail-mealItem.component';
