--- src/components/meal/mealItem/index.ts
@@ -1,2 +1,0 @@
-export { default } from './mealItem.component';
-export * from './mealItem.component';
